Fixed Toolbar error
Was receiving an illegal state exception due to toolbar and actionbar
discrepancy. To fix, I removed the android:windowActionBar>false<
inside the styles xml and added windowActionBar>false<, a more L
friendly declaration to remove the action bar. Next I set the gradle
target version to 21 and the minimum version to 17.